requirements.txt
cleo==0.3.0
Inject==3.3.1
mock==1.3.0
nose==1.3.7
pbr==1.6.0
pylev==1.3.0
PyYAML==3.11
six==1.9.0
SQLAlchemy==1.0.8
tornado==4.2.1
wheel==0.24.0
